# Faster packing 