--- a/.gitignore
+++ b/.gitignore
@@ -63,6 +63,7 @@ htmlcov/*
 junit.xml
 coverage.xml
 .pytest_cache/
+tests/util/fastpack_*.txt
 
 # Build and docs folder/files
 build/*