--- a/src/finn/util/data_packing.py
+++ b/src/finn/util/data_packing.py
@@ -103,7 +103,7 @@ def array2hexstring(array, dtype, pad_to_nbits, prefix="0x", reverse=False, use_
     # TODO: Expand this to cover more cases
     if use_fastpack and dtype == DataType["BINARY"] and prefix == "0x":
         output_string = ct.create_string_buffer(ceil(pad_to_nbits / 4) + 4)
-        success = fastpack.array_to_hexstring_binary(array, array.size, pad_to_nbits, output_string)
+        success = fastpack.array_to_hexstring_binary(np.asarray(array, order='C'), array.size, pad_to_nbits, output_string)
         assert success, f"Could not convert array {array} with datatype {dtype} to hexstring!"
         return output_string.value.decode("utf-8")